requirements.txt
Django==2.2.1
psycopg2-binary==2.7.5
pytz==2018.5
whitenoise==4.1
pillow==5.3.0
django-widget-tweaks==1.4.5
channels
django-friendship
channels_redis



